moved pbject kind to a different folder
