+++ integrations/jira/object_kind.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class ObjectKind(StrEnum):
+    PROJECT = "project"
+    ISSUE = "issue"
+    USER = "user"
+    TEAM = "team"
